+++ index.ts
@@ -1,1 +1,5 @@
 export { configure } from './configure.js'
+export { observeAxios } from './src/axios.js'
+export { reportError } from './src/reportErrors.js'
+import APIToolkitMiddleware from './src/middleware/apitoolkit_middleware.js'
+export default APIToolkitMiddleware
